+++ mlflow/store/job/sqlalchemy_store.py
@@ -41,7 +41,7 @@ class SqlAlchemyJobStore(AbstractJobStore):
             SessionMaker, self.db_type
         )
 
-    def create_job(self, function_fullname: str, params: str, timeout: int | None = None) -> Job:
+    def create_job(self, function_fullname: str, params: str, timeout: float | None = None) -> Job:
         """
         Create a new job with the specified function and parameters.
